+++ src/client.js
@@ -1,5 +1,5 @@
-// import * as Sentry from "@sentry/browser";
-// import { Integrations } from "@sentry/tracing";
+import * as Sentry from "@sentry/browser";
+import { Integrations } from "@sentry/tracing";
 import * as sapper from "@sapper/app";
 import { injectVendor } from "@helpers/vendor";
 
@@ -13,13 +13,16 @@ if (process.env.ENABLE_ERUDA === "true") {
   });
 }
 
-// Sentry.init({
-//   dsn: "https://1a013830891b487abc0e68134ca44237@o1000727.ingest.sentry.io/5960158",
-//   release: "kiluan@1.2.2",
-//   integrations: [new Integrations.BrowserTracing()],
-//   tracesSampleRate: 1.0,
-//   environment: process.env.NODE_ENV === "development" ? "development" : "production",
-// });
+if (process.env.SENTRY_ENABLED === "true") {
+  Sentry.init({
+    dsn: process.env.SENTRY_DSN,
+    release: process.env.npm_package_version,
+    integrations: [new Integrations.BrowserTracing()],
+    tracesSampleRate: 1.0,
+    environment:
+      process.env.NODE_ENV === "development" ? "development" : "production",
+  });
+}
 
 sapper.start({
   target: document.querySelector("#sapper"),