+++ src/server/utils/log.js
@@ -1,4 +1,5 @@
 import loggertdr from "@server/utils/loggertdr";
+import logger from "@server/utils/logger";
 import { logProcess, logStartProcess, logEndProcess } from "@constants/logger";
 
 const moment = require("moment-timezone");
@@ -6,26 +7,31 @@ moment.tz.setDefault("Asia/Jakarta");
 const mapMetadata = new Map();
 
 const intruder = "intruders";
-const appname = "kiluan";
+const appname = "sempu";
 
-export const ts = (rid, tid, tag, path, msg) => {
+export const ts = (rid, tid, tag, path, msg, logType = "") => {
   let logTime = new Date();
   let messageLog = {
     requestId: rid,
     transId: tid || intruder,
     tag: tag,
+    logType,
     domain: path.domain,
     path: path.pathname,
     logTime: logTime.toISOString(),
     datas: msg,
   };
 
-  loggertdr.info(JSON.stringify(messageLog));
+  if (logType === "SYS") {
+    logger.info(JSON.stringify(messageLog));
+  } else {
+    loggertdr.info(JSON.stringify(messageLog));
+  }
 
   mapMetadata.set(tid, logTime);
 };
 
-export const te = (rid, tid, tag, path, msg) => {
+export const te = (rid, tid, tag, path, msg, logType = "") => {
   let logStartTime = mapMetadata.get(tid);
   let logTime = new Date();
   let diff = Math.abs(logTime - logStartTime); // difference in milliseconds
@@ -33,6 +39,7 @@ export const te = (rid, tid, tag, path, msg) => {
     requestId: rid,
     transId: tid || intruder,
     tag: tag,
+    logType,
     domain: path.domain,
     path: path.pathname,
     logTime: logTime.toISOString(),
@@ -40,12 +47,16 @@ export const te = (rid, tid, tag, path, msg) => {
     datas: msg,
   };
 
-  loggertdr.info(JSON.stringify(messageLog));
+  if (logType === "SYS") {
+    logger.info(JSON.stringify(messageLog));
+  } else {
+    loggertdr.info(JSON.stringify(messageLog));
+  }
 
   mapMetadata.delete(tid);
 };
 
-export const tdr = (rid, tid, path, msg) => {
+export const tdr = (rid, tid, path, msg, logType = "") => {
   const logStartTime = mapMetadata.get(tid);
   const logTime = new Date();
   const diff = Math.abs(logTime - logStartTime); // difference in milliseconds
@@ -54,6 +65,7 @@ export const tdr = (rid, tid, path, msg) => {
     transId: tid || intruder,
     port: process.env.PORT || 3010,
     logTime: logTime.toISOString(),
+    logType,
     app: appname,
     ver: "v1",
     domain: path.domain,
@@ -64,21 +76,30 @@ export const tdr = (rid, tid, path, msg) => {
     resp: msg.response,
   };
 
-  loggertdr.info(JSON.stringify(messageLog));
+  if (logType === "SYS") {
+    logger.info(JSON.stringify(messageLog));
+  } else {
+    loggertdr.info(JSON.stringify(messageLog));
+  }
 };
 
-export const tp = (rid, tid, pname, msg) => {
+export const tp = (rid, tid, pname, msg, logType = "") => {
   let logTime = new Date();
   let messageLog = {
     requestId: rid,
     transId: tid,
     tag: logProcess,
+    logType,
     processName: pname,
     datas: msg,
     logTime: logTime.toISOString(),
   };
 
-  loggertdr.info(JSON.stringify(messageLog));
+  if (logType === "SYS") {
+    logger.info(JSON.stringify(messageLog));
+  } else {
+    loggertdr.info(JSON.stringify(messageLog));
+  }
 };
 
 export const tps = (rid, tid, pname, msg) => {